Removed integration tests from unit test classes
