+++ src/test/java/com/quinnshultz/opal/test/bean/PasswordAccountEncryptionIT.java
@@ -66,6 +66,45 @@ public class PasswordAccountEncryptionIT {
 		assertNotNull(key);
 	}
 	
+	/**
+	 * Tests that the setEncryptedData() method does not throw an Exception
+	 * @throws NoSuchAlgorithmException Likely because the algorithm was mismatched somewhere, "AES" may be hardcoded
+	 */
+	@Test
+	public void testSetEncryptedData() throws NoSuchAlgorithmException {
+		SecretKey key = DataEncrypter.generateKey();
+		String encryptedData = DataEncrypter.encryptString(METHOD_PASSWORD, key.getEncoded());
+		try {
+			// TODO: Get bytes may be a nondeterministic method, this may need to be corrected in multiple locations
+			encrypter.setEncryptedData(encryptedData.getBytes());
+		} catch (Exception e) {
+			fail("Caught an Exception when executing setEncryptedData()");
+		}
+	}
 	
+	// TODO: Test getEncryptedData() after setEncryptedData()
+	
+	// TODO: Test that the getData() method returns expected results uninitialized
+	
+	/**
+	 * Tests that the setData() method does not throw an Exception
+	 */
+	@Test
+	public void testSetData() {
+		try {
+			encrypter.setData(METHOD_PASSWORD, key);
+		} catch (Exception e) {
+			fail("Caught an Exception when executing setData()");
+		}
+	}
+	
+	/**
+	 * Tests that expected results are returned by getData() after the the setData() method is called
+	 */
+	@Test
+	public void testGetDataAfterSetData() {
+		encrypter.setData(METHOD_PASSWORD, key);
+		assertEquals(encrypter.getData(key), METHOD_PASSWORD);
+	}
 
 }
