+++ src/test/java/com/quinnshultz/opal/test/unit/bean/PADefaultConstructorTest.java
@@ -21,22 +21,19 @@ import static org.junit.Assert.fail;
 
 import java.security.NoSuchAlgorithmException;
 
-import javax.crypto.SecretKey;
 
 import org.junit.Before;
 import org.junit.Test;
 
-import com.quinnshultz.opal.bean.OpalUser;
 import com.quinnshultz.opal.bean.PasswordAccount;
-import com.quinnshultz.opal.util.DataEncrypter;
+
 
 /**
  * Tests class PasswordAccount
  * @author Quinn Shultz
  */
 public class PADefaultConstructorTest {
-	
-	private final String CONSTRUCTOR_FULL_NAME = "John Doe";
+
 	
 	private final int METHOD_ID = 22;
 	private final String METHOD_URL = "https://www.deezer.com/us/login";
@@ -46,10 +43,7 @@ public class PADefaultConstructorTest {
 	private final String METHOD_PASSWORD = "6A6l$!nEe4ifD0@E";
 	private final String METHOD_NOTES = "A music streaming service.";
 	
-	private byte[] key;
-	
 	private PasswordAccount encrypter;
-	private OpalUser account;
 
 	/**
 	 * Constructs a new PasswordAccount Object
@@ -57,8 +51,6 @@ public class PADefaultConstructorTest {
 	 */
 	@Before
 	public void setUp() throws NoSuchAlgorithmException {
-		account = new OpalUser(METHOD_USERNAME, METHOD_PASSWORD, CONSTRUCTOR_FULL_NAME);
-		key = account.getSerializedKey();
 		encrypter = new PasswordAccount();
 	}
 	
@@ -148,46 +140,13 @@ public class PADefaultConstructorTest {
 		}
 	}
 	
-	/**
-	 * Tests that the setEncryptedData() method does not throw an Exception
-	 * @throws NoSuchAlgorithmException Likely because the algorithm was mismatched somewhere, "AES" may be hardcoded
-	 */
-	@Test
-	public void testSetEncryptedData() throws NoSuchAlgorithmException {
-		SecretKey key = DataEncrypter.generateKey();
-		String encryptedData = DataEncrypter.encryptString(METHOD_PASSWORD, key.getEncoded());
-		try {
-			// TODO: Get bytes may be a nondeterministic method, this may need to be corrected in multiple locations
-			encrypter.setEncryptedData(encryptedData.getBytes());
-		} catch (Exception e) {
-			fail("Caught an Exception when executing setEncryptedData()");
-		}
-	}
+
 	
 	// TODO: Test getEncryptedData() after setEncryptedData()
 	
 	// TODO: Test that the getData() method returns expected results uninitialized
 	
-	/**
-	 * Tests that the setData() method does not throw an Exception
-	 */
-	@Test
-	public void testSetData() {
-		try {
-			encrypter.setData(METHOD_PASSWORD, key);
-		} catch (Exception e) {
-			fail("Caught an Exception when executing setData()");
-		}
-	}
-	
-	/**
-	 * Tests that expected results are returned by getData() after the the setData() method is called
-	 */
-	@Test
-	public void testGetDataAfterSetData() {
-		encrypter.setData(METHOD_PASSWORD, key);
-		assertEquals(encrypter.getData(key), METHOD_PASSWORD);
-	}
+
 	
 	/**
 	 * Tests that the getName() method returns expected results and throws no Exception
